requirements.txt
bedrock-agentcore>=1.2.0
strands-agents>=1.24.0
strands-agents-tools>=0.2.19
boto3>=1.42.0
pandas
numpy
matplotlib
